Add pair_range to use std::pair<Iterator> in range based for loop.
normale the loop only iterators over the pair itself.
This is not terribly useful, when one wants to iterator over the range
of the iterators within the pair.
This is particularly annoying with algorithms like equal_range, which return pairs.

Note: might need to add some const overloads and noexcept